refactor: move reservability functions
